backend/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn==0.24.0

# Database (MongoDB)
motor==3.3.2
pymongo==4.5.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0
